* C++ seems to be serious on type sizes (I want a system, where I have the same error, dammit)

+++ genet/libs/cudd/dddmp/dddmpNodeBdd.c
@@ -343,7 +343,7 @@ RemoveFromUniqueRecurBdd (
 
     nodelist = subtable->nodelist;
 
-    pos = ddHash ((unsigned int) cuddT (f),(unsigned int) cuddE (f), subtable->shift);
+    pos = ddHash ((unsigned long) cuddT (f),(unsigned long) cuddE (f), subtable->shift);
     node = nodelist[pos];
     last = NULL;
     while (node != sentinel) {
@@ -421,7 +421,7 @@ RestoreInUniqueRecurBdd (
   nodelist = subtable->nodelist;
 
   // \TODO: check definition of ddHash (include/CuddInt.h) for type sizes
-  pos = ddHash ((unsigned int) cuddT (f),(unsigned int) cuddE (f), subtable->shift);
+  pos = ddHash ((unsigned long) cuddT (f),(unsigned long) cuddE (f), subtable->shift);
 
 #ifdef DDDMP_DEBUG
   /* verify uniqueness to avoid duplicate nodes in unique table */
